+++ blocks/top-topics/top-topics.css
@@ -1,0 +1,254 @@
+.top-topics {
+    width: 100%;
+    display: flex;
+    flex-direction: column;
+}
+
+.top-topics .tabs-list {
+    width: 100%;
+    padding: 0 0 33px;
+}
+
+.top-topics .tabs-tab {
+    width: 100%;
+}
+
+.top-topics .tabs-list button {
+    display: block;
+    width: 100%;
+    margin: 0;
+    padding: 10px 50px;
+    text-align: left;
+    cursor: pointer;
+    position: relative;
+    border: none;
+}
+
+.top-topics .tabs-list button::after {
+    content: '';
+    position: absolute;
+    right: 0;
+    top: 50%;
+    transform: translateY(-50%);
+    background: url('../../icons/chevron-right-white.svg') center center no-repeat;
+    width: 24px;
+    height: 24px;
+    opacity: 0;
+}
+
+.top-topics .tabs-list button[aria-selected="true"]::after {
+    opacity: 1;
+}
+
+/* .top-topics .tabs-list button[aria-selected="true"] {
+    cursor: pointer;
+} */
+
+/* .top-topics .tabs-list button[aria-selected="false"]:hover,
+.top-topics .tabs-list button[aria-selected="false"]:focus {
+    background-color: var(--dark-color);
+} */
+
+.top-topics .tabs-panel[aria-hidden="false"] {
+    /* margin-top: -1px; */
+    padding: 0 16px;
+    /* height: var(--tab-height); */
+    /* transition: height .5s ease-in; */
+}
+
+.top-topics .tabs-panel[aria-hidden="true"] {
+    height: 0;
+    overflow: hidden;
+    transition: height .5s ease-in;
+}
+
+.top-topics a {
+    display: inline-block;
+    font: 400 18px var(--body-font-family);
+    color: #000;
+}
+
+.top-topics .tabs-panel ul > li > a {
+    text-wrap: wrap;
+}
+
+.top-topics .tabs-panel a:any-link {
+    padding: 0;
+    margin-bottom: 24px;
+    color: var(--primary-orange-600);
+    text-decoration: underline;
+}
+
+.top-topics .tabs-panel ul {
+    column-count: 2;
+    column-gap: 10vw;
+    list-style: none;
+}
+
+.top-topics p {
+    font-size: var(--body-font-size-m);
+    color: var(--primary-orange-600);
+    font-style: normal;
+    font-weight: var(--font-weight-bold);
+    line-height: 110%;
+    cursor: pointer;
+    /* transition: all 300ms; */
+    position: relative;
+    margin: 0;
+}
+
+.tabs-tab[aria-selected="true"] > p {
+    color: var(--orange);
+}
+
+.top-topics .tabs-panel > div p {
+    display: none;
+}
+
+.tabs-container h2 {
+    background-color: white;
+    margin: 0;
+    color: var(--orange);
+    padding: 38px 0 0 3vw;
+    width: 100%;
+    font: 650 34px var(--body-font-family);
+    padding-bottom: 30px;
+}
+
+.section.top-topics-container div.default-content-wrapper:last-child {
+    display: flex;
+    justify-content: flex-start;
+    background-color: white;
+    width: 100%;
+    align-items: center;
+    font: 400 18px var(--body-font-family);
+}
+
+.section.top-topics-container div.default-content-wrapper:last-child img {
+    height: auto;
+}
+
+.section.top-topics-container div.default-content-wrapper:last-child p {
+    padding-top : 5vh;
+    padding-bottom: 8vh;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    gap: 2.5vw;
+    flex-direction: row;
+    border-top: 2.5px solid #000;
+    padding-left: 0;
+    margin-left: 3vw;
+}
+
+.section.top-topics-container div.default-content-wrapper:last-child p:last-child picture {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
+
+@media (width >= 900px) {
+    .top-topics {
+        flex-direction: row;
+        max-width: 1440px;
+    } 
+
+    /* .top-topics p {
+        padding-right: 160px;
+        padding-left: 0;
+    } */
+
+    .top-topics .tabs-list {
+        width: 100%;
+        max-width: 350px;
+        background-color: var(--light-color);
+        padding: 50px 0;
+    }
+
+    .top-topics .tabs-list button {
+        background-color: var(--light-color);
+    }
+
+    .top-topics-container h2 {
+        background-color: rgb(238 238 238);
+        width: 40%;
+    }
+
+    .top-topics .tabs-panel[aria-hidden="false"] {
+        width: 100%;
+        max-width: none;
+        padding: 50px;
+        min-height: 500px;
+    }
+
+    .top-topics .tabs-panel > div p {
+        display: block;
+        color: var(--primary-orange-600);
+        font-size: var(--heading-font-size-xl);
+        font-style: normal;
+        font-weight: var(--font-weight-bold);
+        line-height: 120%;
+        margin-bottom: 50px;
+    }  
+
+    .top-topics .tabs-panel[aria-hidden="true"] {
+        display: none;
+    }
+
+    .top-topics .tabs-panel ul {
+        margin: 0;
+        padding: 0;
+    }
+
+    .top-topics .tabs-list button::after {
+        content: '';
+        position: absolute;
+        right: 8px;
+        top: 50%;
+        transform: translateY(-50%); 
+        background: url('../../icons/chevron-right-white.svg') center center no-repeat;
+        width: 24px;
+        height: 24px;
+    }
+
+    .top-topics .tabs-list button[aria-selected="true"] {
+        background-color: var(--primary-orange-600);
+    }
+
+    .top-topics .tabs-list button[aria-selected="true"] p {
+        color: var(--neutral-white);
+    }
+
+    /* .top-topics .tabs-list button[aria-selected="true"]::after {
+        transform: translateY(-50%) rotate(-360deg);
+    } */
+
+    .section.top-topics-container div.default-content-wrapper:last-child {
+        display: flex;
+        justify-content: flex-start;
+        background-color: var(--light-color);
+        width: 40%;
+        align-items: center;
+        font: 400 18px var(--body-font-family);
+        margin-left: 0;
+    }
+
+    .section.top-topics-container div.default-content-wrapper:last-child p {
+        padding-top : 5vh;
+        padding-bottom: 8vh;
+        display: flex;
+        justify-content: center;
+        align-items: center;
+        gap: 2.5vw;
+        flex-direction: row;
+        border-top: 2.5px solid #000;
+        padding-left: 0;
+        margin-left: 3vw;
+    }
+
+    .section.top-topics-container div.default-content-wrapper:last-child p:last-child picture {
+        display: flex;
+        justify-content: center;
+        align-items: center;
+    }
+}
